+++ lib/wallet/src/Cardano/DB/Sqlite.hs
@@ -41,7 +41,6 @@ module Cardano.DB.Sqlite
     , dbChunked
     , dbChunkedFor
     , dbChunked'
-    , handleConstraint
 
       -- * Logging
     , DBLog (..)
@@ -83,9 +82,6 @@ import Control.Lens
 import Control.Monad
     ( join
     , void
-    )
-import Control.Monad.IO.Unlift
-    ( MonadUnliftIO (..)
     )
 import Control.Monad.Logger
     ( LogLevel (..)
@@ -155,8 +151,7 @@ import Database.Persist.Sqlite
     , wrapConnection
     )
 import Database.Sqlite
-    ( Error (ErrorConstraint)
-    , SqliteException (SqliteException)
+    ( SqliteException (SqliteException)
     )
 import Fmt
     ( Buildable (..)
@@ -181,7 +176,6 @@ import UnliftIO.Compat
     )
 import UnliftIO.Exception
     ( bracket
-    , handleJust
     , tryJust
     )
 import UnliftIO.MVar
@@ -204,15 +198,6 @@ import qualified Database.Sqlite as Sqlite
 newtype SqliteContext = SqliteContext
     { runQuery :: forall a. SqlPersistT IO a -> IO a
     }
-
--- | Run an action, and convert any Sqlite constraints exception into the given
--- error result. No other exceptions are handled.
-handleConstraint :: MonadUnliftIO m => e -> m a -> m (Either e a)
-handleConstraint e = handleJust select handler . fmap Right
-  where
-    select (SqliteException ErrorConstraint _ _) = Just ()
-    select _ = Nothing
-    handler = const . pure . Left $ e
 
 newInMemorySqliteContext
     :: Tracer IO DBLog
